Support comma separated tags everywhere
This should remove some confusion on comma vs. space separated tags.
The prompts now tell you to separate tags with commas and it errors out
if you try to have a space in a tag. While it had been possible to have
spaces in tags, it was really confusing to see them printed and when
they were added, it was an accident.

This also updates the `--tag` option so that if you pass `--tag a,b,c`,
it will turn into 3 tags. This is the same behavior as `--tag a --tag b
--tag c`, but shorter and without the need to parse tags in programs
calling `mix nerves_hub.device create`.