Support CA Certificate serial as hex
The web UI was recently changed to display certificate serial numbers as
hex. This updates `nerves_hub.ca_certificate` functions to also display
the serial as hex as well. It also adjusts `nerves_hub.ca_certificate
unregister` to support using the hex or interger representation of the
serial as well